move swagger options to swagger.ts
